+++ app/src/main/java/com/example/jwstimager/MainActivity.kt
@@ -1,6 +1,5 @@
 package com.example.jwstimager
 
-import android.content.ContentValues.TAG
 import android.content.Intent
 import android.graphics.Bitmap
 import android.net.Uri
@@ -13,10 +12,10 @@ import androidx.activity.compose.setContent
 import androidx.compose.animation.AnimatedVisibility
 import androidx.compose.animation.animateContentSize
 import androidx.compose.foundation.ExperimentalFoundationApi
-import androidx.compose.foundation.Image
-import androidx.compose.foundation.border
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.*
+//import androidx.compose.foundation.layout.BoxScopeInstance.align
+//import androidx.compose.foundation.layout.ColumnScopeInstance.align
 //import androidx.compose.foundation.layout.BoxScopeInstance.align
 import androidx.compose.foundation.lazy.GridCells
 import androidx.compose.foundation.lazy.LazyColumn
@@ -32,15 +31,13 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.vector.ImageVector
 import androidx.compose.ui.platform.LocalContext
-import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
 import androidx.core.graphics.drawable.toBitmap
 import coil.ImageLoader
 import coil.compose.AsyncImage
 import coil.request.ImageRequest
-import com.example.jwstimager.FeedAPI.BASE_URL
 import com.example.jwstimager.model.Feed
+import com.example.jwstimager.model.entry.Entry
 import com.example.jwstimager.ui.theme.JWSTimagerTheme
 import kotlinx.coroutines.launch
 
@@ -53,11 +50,6 @@ import java.io.ByteArrayOutputStream
 import java.lang.IndexOutOfBoundsException
 import java.lang.NullPointerException
 import java.util.ArrayList
-
-
-
-
-
 
 class MainActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
@@ -146,6 +138,54 @@ class MainActivity : ComponentActivity() {
     }
 }
 
+//
+//
+@Composable
+fun NewsCard(post : Post){
+
+    Box(Modifier.fillMaxWidth()){
+        Column(){
+            Surface(modifier = Modifier
+                .height(100.dp)
+                .padding(3.dp)
+                .fillMaxWidth(),
+                color = MaterialTheme.colorScheme.primary
+            ) {
+                Column(){
+                    Row(){
+                        Column(){
+                            //thumbnail
+                            coil.compose.AsyncImage(model = post.thumbnailURL,
+                                contentDescription = post.title,
+                                modifier = Modifier
+                            )
+                        }
+                        Column(){
+                            //title <-link
+                            Text(
+                                text = post.title
+                            )
+                        }
+                    }
+                    Row(){
+                        //content
+                        Text(
+                            text = post.author
+                        )
+                    }
+                    Row(){
+                        //date
+                        Text(
+                        text = post.date_updated
+                        )
+                    }
+
+
+                }
+            }
+        }
+    }
+}
 
 
 
@@ -274,6 +314,8 @@ fun ImageCard(image: ImageData) {
         }
 }
 
+//
+//
 @Composable
 fun GalleryImageCard(image: ImageData) {
 
@@ -291,8 +333,6 @@ fun GalleryImageCard(image: ImageData) {
                 //toggle is expanded by clicking on the image
                 .clickable { isSelected = !isSelected }
         )
-
-
         Box(
             Modifier
                 .fillMaxSize()
@@ -318,108 +358,33 @@ fun GalleryImageCard(image: ImageData) {
     }
 }
 
-data class AboutData(
-    val name: String,
-    val aboutText: String,
-    val email: String,
-    val linkedin: String
-)
-
-@Composable
-fun AboutCard(aboutEntry: AboutData) {
-    Box(
-        modifier = Modifier
-            .padding(10.dp)
-            .border(1.5.dp, MaterialTheme.colorScheme.primary)
-    )
-    {
-        Column() {
-            Text(text = aboutEntry.name,
-                color = MaterialTheme.colorScheme.secondary,
-                modifier = Modifier.padding(10.dp))
-            Text(text = aboutEntry.aboutText,
-                color = MaterialTheme.colorScheme.secondary,
-                modifier = Modifier.padding(10.dp))
-            Text(text = aboutEntry.email,
-                color = MaterialTheme.colorScheme.secondary,
-                modifier = Modifier.padding(10.dp))
-            Text(text = aboutEntry.linkedin,
-                color = MaterialTheme.colorScheme.secondary,
-                modifier = Modifier.padding(10.dp))
-        }
-    }
-}
 
 
-@Composable
-fun AboutPage() {
 
-    val aboutEntryList = listOf<AboutData>(
-        AboutData(
-            name = androidx.compose.ui.res.stringResource(R.string.nameAndrew),
-            aboutText = androidx.compose.ui.res.stringResource(R.string.about_andrew),
-            email = androidx.compose.ui.res.stringResource(R.string.andrew_email),
-            linkedin = androidx.compose.ui.res.stringResource(R.string.andrew_linkedin),
-        ),
-        AboutData(
-            name = androidx.compose.ui.res.stringResource(R.string.nameDeclan),
-            aboutText = androidx.compose.ui.res.stringResource(R.string.about_declan),
-            email = androidx.compose.ui.res.stringResource(R.string.declan_email),
-            linkedin = androidx.compose.ui.res.stringResource(R.string.declan_linkedin),
-        ),
-        AboutData(
-            name = androidx.compose.ui.res.stringResource(R.string.nameEmmanuel),
-            aboutText = androidx.compose.ui.res.stringResource(R.string.about_emmanuel),
-            email = androidx.compose.ui.res.stringResource(R.string.emmanuel_email),
-            linkedin = androidx.compose.ui.res.stringResource(R.string.emmanuel_linkedin),
-        ),
-
-        AboutData(
-            name = androidx.compose.ui.res.stringResource(R.string.nameJadrien),
-            aboutText = androidx.compose.ui.res.stringResource(R.string.about_jadrien),
-            email = androidx.compose.ui.res.stringResource(R.string.jadrien_email),
-            linkedin = androidx.compose.ui.res.stringResource(R.string.jadrien_linkedin),
-        )
-    )
-    Column() {
-        Text(
-            text = "About Us",
-            color = MaterialTheme.colorScheme.secondary,
-            style = MaterialTheme.typography.titleMedium,
-            modifier = Modifier
-                .padding(10.dp)
-                .align(Alignment.CenterHorizontally)
-
-        )
-        LazyColumn {
-            items(aboutEntryList) { entry ->
-                AboutCard(entry)
-            }
-
-        }
-
-        // Andrew
-
-        // Declan
-
-        // Emmanuel
-
-        // Jadrien
-
-    }
-}
-
-
+///* ------------------------------Scrolling Lists-------------------------------------- *///
 //
 //
 @Composable
-fun ScrollingList(imageList: List<ImageData>) {
+fun ScrollingImageList(imageList: List<ImageData>) {
     LazyColumn {
         items(imageList) { image ->
             ImageCard(image)
         }
     }
 }
+
+
+//
+//
+@Composable
+fun ScrollingNewsList(postList: Post) {
+    LazyColumn {
+        items(postList) { post ->
+            NewsCard(post)
+        }
+    }
+}
+
 
 
 //
@@ -438,87 +403,3 @@ internal fun ScrollingGridList(imageList: List<ImageData>) {
     )
 
 }
-
-
-/*
-//
-//
-@Composable
-fun rememberLazyGridState(
-initialFirstVisibleItemIndex: Int = 0,
-initialFirstVisibleItemScrollOffset: Int = 0
-): LazyGridState {}
-*/
-
-/*
-//
-//
-@Composable
-fun DropdownMenu() {
-var expanded by remember { mutableStateOf(false)}
-val list = listOf("List","Grid","News","About")
-var selectedItem by remember { mutableStateOf( "")}
-var textFiledSize by remember { mutableStateOf(Size.Zero)}
-val icon = if (expanded){
-    Icons.Filled.KeyboardArrowUp
-}else {
-    Icons.Filled.KeyboardArrowDown
-}
-
-Column(modifier = Modifier.padding(20.dp)) {
-
-
-}
-
-}
-*/
-
-
-
-
-
-@Composable
-fun TitleBar() {
-
-    Row(modifier = Modifier.padding(all = 8.dp),
-        verticalAlignment = Alignment.CenterVertically) {
-        Image(
-            painter = painterResource(id = R.mipmap.ic_launcher_foreground),
-            contentDescription = "logo",
-            modifier = Modifier.size(70.dp, 70.dp)
-        )
-        //Column() {
-        Text("JWSTimager",
-            color = Color.White,
-            //textAlign = TextAlign.Justify,
-            fontSize = 32.sp)
-        // style = MaterialTheme.typography.headlineMedium
-        //}
-
-    }
-}
-
-
-
-
-
-/*
-@Composable
-fun DefaultPreview() {
-    JWSTimagerTheme {
-        Surface(
-            modifier = Modifier.fillMaxSize(),
-            color = MaterialTheme.colorScheme.primary
-        ) {
-            Column {
-                TitleBar()
-                //ScrollingList(imageList = SampleData.sampleImageList)
-                ScrollingGridList(imageList = SampleData.sampleImageList)
-
-            }
-        }
-
-
-    }
-}
-*/
