+++ app/src/main/java/com/example/jwstimager/TitleBar.kt
@@ -1,0 +1,36 @@
+package com.example.jwstimager
+
+import androidx.compose.foundation.Image
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+
+
+@Composable
+fun TitleBar() {
+
+    Row(modifier = Modifier.padding(all = 8.dp),
+        verticalAlignment = Alignment.CenterVertically) {
+        Image(
+            painter = painterResource(id = R.mipmap.ic_launcher_foreground),
+            contentDescription = "logo",
+            modifier = Modifier.size(70.dp, 70.dp)
+        )
+        //Column() {
+        Text("JWSTimager",
+            color = Color.White,
+            //textAlign = TextAlign.Justify,
+            fontSize = 32.sp)
+        // style = MaterialTheme.typography.headlineMedium
+        //}
+
+    }
+}
